+++ android/src/ti/map/Shape/IShape.java
@@ -6,11 +6,5 @@ import ti.map.TiMarker;
 import com.google.android.gms.maps.model.Marker;
 
 public interface IShape {
-
-	public void setTiMarker(TiMarker marker);
-	
-	public TiMarker getTiMarker();
-	public Marker getMarker();
-	public AnnotationProxy getAnnotation();
 	
 }